# target/generated-sources/openapi/src/main/java/com/gila/ecommerce/api/OrdersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gila.ecommerce.api;

import com.gila.ecommerce.dto.OrderDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-10T21:53:01.882557200-06:00[GMT-06:00]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Orders", description = "the Orders API")
public interface OrdersApi {

    /**
     * POST /api/v1/orders/checkout : Process checkout for the current cart (Customer only)
     *
     * @return Order checked out and payment authorized successfully (status code 200)
     *         or Empty cart or stock validation failed (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "checkout",
        summary = "Process checkout for the current cart (Customer only)",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Order checked out and payment authorized successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Empty cart or stock validation failed"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/orders/checkout",
        produces = { "application/json" }
    )
    
    ResponseEntity<OrderDto> checkout(
        
    );


    /**
     * DELETE /api/v1/orders/clear : Truncate database orders and reset system inventory defaults (Admin only)
     *
     * @return Orders database reset completed successfully (status code 204)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "clearOrders",
        summary = "Truncate database orders and reset system inventory defaults (Admin only)",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Orders database reset completed successfully"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/orders/clear"
    )
    
    ResponseEntity<Void> clearOrders(
        
    );

}