# target/generated-sources/openapi/src/main/java/com/gila/ecommerce/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gila.ecommerce.api;

import com.gila.ecommerce.dto.ImportProducts202Response;
import com.gila.ecommerce.dto.ProductDto;
import com.gila.ecommerce.dto.ProductImportStatusDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-10T21:53:01.882557200-06:00[GMT-06:00]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Products", description = "the Products API")
public interface ProductsApi {

    /**
     * POST /api/v1/products : Create a new product (Admin only)
     *
     * @param productDto  (required)
     * @return Product successfully created (status code 201)
     *         or Invalid product details supplied (status code 400)
     *         or Forbidden operation (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create a new product (Admin only)",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Product successfully created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid product details supplied"),
            @ApiResponse(responseCode = "403", description = "Forbidden operation (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductDto> createProduct(
        @Parameter(name = "ProductDto", description = "", required = true) @Valid @RequestBody ProductDto productDto
    );


    /**
     * DELETE /api/v1/products/{id} : Delete a product (Admin only)
     *
     * @param id  (required)
     * @return Product deleted successfully (status code 204)
     *         or Product not found (status code 404)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Delete a product (Admin only)",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Product not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/products/{id}"
    )
    
    ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/v1/products/import/status/{taskId} : Check status and validation reports of a CSV import task (Admin only)
     *
     * @param taskId  (required)
     * @return Task import status report retrieved (status code 200)
     *         or Import task not found (status code 404)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "getImportStatus",
        summary = "Check status and validation reports of a CSV import task (Admin only)",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task import status report retrieved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductImportStatusDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Import task not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/products/import/status/{taskId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProductImportStatusDto> getImportStatus(
        @Parameter(name = "taskId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("taskId") UUID taskId
    );


    /**
     * GET /api/v1/products/{id} : Get details of a single product
     *
     * @param id  (required)
     * @return Product details retrieved (status code 200)
     *         or Product not found (status code 404)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Get details of a single product",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product details retrieved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/products/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProductDto> getProductById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/v1/products : Retrieve products with optional filter/search
     *
     * @param query Search term matching product name or description (optional)
     * @param category Filter products by category (optional)
     * @param page Page number (0-indexed) (optional, default to 0)
     * @param size Page size limit (optional, default to 10)
     * @return List of products matching criteria (status code 200)
     */
    @Operation(
        operationId = "getProducts",
        summary = "Retrieve products with optional filter/search",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of products matching criteria", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/products",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ProductDto>> getProducts(
        @Parameter(name = "query", description = "Search term matching product name or description", in = ParameterIn.QUERY) @Valid @RequestParam(value = "query", required = false) String query,
        @Parameter(name = "category", description = "Filter products by category", in = ParameterIn.QUERY) @Valid @RequestParam(value = "category", required = false) String category,
        @Parameter(name = "page", description = "Page number (0-indexed)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "Page size limit", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
    );


    /**
     * POST /api/v1/products/import : Asynchronously import products from a CSV file (Admin only)
     *
     * @param file The CSV file to upload containing product records (optional)
     * @return CSV import task accepted and queued for Kafka processing (status code 202)
     *         or Invalid file format or missing CSV headers (status code 400)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "importProducts",
        summary = "Asynchronously import products from a CSV file (Admin only)",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "202", description = "CSV import task accepted and queued for Kafka processing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ImportProducts202Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid file format or missing CSV headers"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/products/import",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    ResponseEntity<ImportProducts202Response> importProducts(
        @Parameter(name = "file", description = "The CSV file to upload containing product records") @RequestPart(value = "file", required = false) MultipartFile file
    );


    /**
     * PUT /api/v1/products/{id} : Update an existing product (Admin only)
     *
     * @param id  (required)
     * @param productDto  (required)
     * @return Product updated successfully (status code 200)
     *         or Product not found (status code 404)
     *         or Forbidden operation (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Update an existing product (Admin only)",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden operation (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductDto> updateProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ProductDto", description = "", required = true) @Valid @RequestBody ProductDto productDto
    );

}